Backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
pymongo==4.7.3
python-dotenv==1.0.1
pdfplumber==0.11.0
gunicorn==22.0.0
